fix some vc6 compile errors
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@3725 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/win32com/Controller.cpp
@@ -42,7 +42,7 @@ extern UINT32 g_raw_dmdy;
 extern char g_fShowWinDMD;
 
 // from ticker.c
-extern void uSleep(const unsigned long long u);
+extern void uSleep(const UINT64 u);
 }
 #include "alias.h"
