+++ src/windows/video.c
@@ -40,7 +40,7 @@ extern int verbose;
 extern struct rc_option win_d3d_opts[];
 
 // from ticker.c
-extern void uSleep(const unsigned long long u);
+extern void uSleep(const UINT64 u);
 
 //============================================================
 //	PARAMETERS
@@ -724,8 +724,8 @@ static void throttle_speed(void)
 		while (curr - target < 0)
 		{
 #ifdef VPINMAME
-			//if((long long)((target - curr)/(ticks_per_sleep_msec*1.1))-1 > 0) // pessimistic estimate of stuff below
-			//	uSleep((unsigned long long)((target - curr)*1000/(ticks_per_sleep_msec*1.1))-1);
+			//if((INT64)((target - curr)/(ticks_per_sleep_msec*1.1))-1 > 0) // pessimistic estimate of stuff below
+			//	uSleep((UINT64)((target - curr)*1000/(ticks_per_sleep_msec*1.1))-1);
 			
 			uSleep((target-curr)*1000/ticks_per_sleep_msec);
 #else
